refactor(add): extract preparation.py (prepare_item + conversion routing) (#623)
Move the per-item preparation/conversion phase out of the 3.5k-line add.py
into a new lower-layer module, portolan_cli/preparation.py. This is the
cleanest extractable seam of add.py (part of the #618 epic).

Moved verbatim (no behavior change):
- prepare_item and its item-level helpers (_create_and_save_item,
  _apply_nodata_defaults_to_bands, _fix_collection_level_asset_hrefs)
- conversion routing: _convert_and_extract_metadata, convert_vector,
  convert_tabular, convert_raster, _handle_cloud_native_vector,
  _warn_if_source_newer
- statistics: _extract_statistics_best_effort, _add_statistics_to_properties
- the item-prep primitives prepare_item depends on (_scan_item_assets and its
  asset-key helpers, _derive_item_id_and_asset_level, _validate_collection_id,
  _pre_validate_geometry, _extract_bbox_wgs84, _cleanup_orphaned_output,
  PreparedItem, the metadata type aliases)

The dependency edge is one-directional (add -> preparation); preparation
imports nothing from add, so there is no cycle. add.py re-exports the names its
finalize/asset-update code and the test-suite still import from it; external.py
now imports _validate_collection_id from preparation.

add.py: 3483 -> 2392 lines. Adds tests/unit/test_preparation.py covering the
conversion routing, force/reconvert semantics, statistics shaping, collection-
level href/key fixing, collection-id validation, and bbox CRS handling.

Gates: import-linter, mypy --strict, ruff, vulture, xenon, pylint
duplicate-code, and the full unit + add integration suites all green.
