+++ portolan_cli/external.py
@@ -29,10 +29,8 @@ from urllib.parse import urlparse
 
 import pystac
 
-from portolan_cli.add import (
-    _save_collection_with_links,
-    _validate_collection_id,
-)
+from portolan_cli.add import _save_collection_with_links
+from portolan_cli.preparation import _validate_collection_id
 from portolan_cli.stac import (
     DEFAULT_LICENSE,
     add_asset_to_collection,
